requirements.txt
cryptography==42.0.5
Flask==3.0.2
qrcode==7.4.2
pillow==10.3.0
opencv-python==4.9.0.80
reportlab==4.2.0
